+++ algorithms/detection.hpp
